refactor: rename files in blocks
